# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API сервиса iAmOkey",
        "description": "Документация сервиса.",
        "contact": {
            "email": "nfs2025@mail.ru"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/accelerometer/{patient_id}": {
            "get": {
                "tags": [
                    "Accelerometer"
                ],
                "summary": "Показатели пациента",
                "description": "Возвращается объект всех показаний по пациенту",
                "operationId": "8aa303b27b0644f8662526b36e840a30",
                "parameters": [
                    {
                        "name": "patient_id",
                        "in": "path",
                        "description": "ID пациента",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
        "/login/": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Авторизация",
                "description": "Авторизация, возвращается объект пользователя",
                "operationId": "15553fe7f5d4333c24767d8b33d03be2",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "path",
                        "description": "Телефон",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/confim/": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Подтверждение по коду",
                "description": "Авторизация, возвращается объект пользователя",
                "operationId": "34f185271f7b75d35c128031e76bd1ac",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "path",
                        "description": "Телефон",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Код",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
        "/calibration/{patient_id}": {
            "get": {
                "tags": [
                    "Calibration"
                ],
                "summary": "Калибровка пациента",
                "description": "Возвращается объект показателей калибровки по пациенту",
                "operationId": "29408b86d39d08abc4aa697bb65c946a",
                "parameters": [
                    {
                        "name": "patient_id",
                        "in": "path",
                        "description": "ID пациента",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/calibration/": {
            "put": {
                "tags": [
                    "Calibration"
                ],
                "summary": "Создать/обновить калибровка",
                "description": "Создать/обновить калибровку, возвращается объект калибровки",
                "operationId": "0467d637b882d5b80c8265c34f03670f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Calibration"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Calibration"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Calibration": {
                "title": "Calibration",
                "description": "Calibration model",
                "properties": {
                    "patient_id": {
                        "title": "patient_id",
                        "description": "ID пациента",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "systolic_1": {
                        "title": "systolic_1",
                        "description": "Систолическое давление 1",
                        "type": "integer",
                        "format": "int64",
                        "example": 120
                    },
                    "systolic_2": {
                        "title": "systolic_2",
                        "description": "Систолическое давление 2",
                        "type": "integer",
                        "format": "int64",
                        "example": 120
                    },
                    "systolic_3": {
                        "title": "systolic_3",
                        "description": "Систолическое давление 3",
                        "type": "integer",
                        "format": "int64",
                        "example": 120
                    },
                    "diastolic_1": {
                        "title": "diastolic_1",
                        "description": "Диастолическое давление 1",
                        "type": "integer",
                        "format": "int64",
                        "example": 80
                    },
                    "diastolic_2": {
                        "title": "diastolic_2",
                        "description": "Диастолическое давление 2",
                        "type": "integer",
                        "format": "int64",
                        "example": 80
                    },
                    "diastolic_3": {
                        "title": "diastolic_3",
                        "description": "Диастолическое давление 3",
                        "type": "integer",
                        "format": "int64",
                        "example": 80
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Calibration"
                }
            },
...
        },
...
    },
...
    "security": [
        []
    ]
}